Modified the type of the constant STEPS from int to double

+++ EcoSorter/EcoSorterProject/EcoSorterProject.h
@@ -10,7 +10,7 @@ private:
 	HANDLE						visionThread;
 
 	double						CVT_CM2PX;
-	int								STEPS;
+	double						STEPS;
 
 public:
 	EcoSorterProject();
